TODO 34 : faire en sorte que le line render cause des dégats x degat tous les x sec (en float) ; also fix little issue

+++ Assets/Scripts/TriggerZoneDamage.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class TriggerZoneDamage : MonoBehaviour
 {
     [SerializeField] string tagToDamage = null;
-    [SerializeField] int damage = 0;
+    [SerializeField] float damage = 0;
 
     private void OnTriggerEnter(Collider other)
     {
